refactor: extract account fetching logic for reuse
- moves account fetching/sorting from Accounts to PlanetCash (parent)
- enables reuse of data in CreateAccount and Accounts
